% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\alias{cluster,tNN,data.frame-method}
\alias{cluster,tNN,matrix-method}
\alias{cluster,tNN,numeric-method}
\title{Data stream clustering with tNN}
\usage{
cluster(x, newdata, ...)
}
\arguments{
\item{x}{a \code{tNN} object. Note that this function canges the original object!}

\item{newdata}{a vector (one observation), or a matrix or data.frame
(each row is an observation).}

\item{...}{further arguments like \code{verbose}.}
}
\value{
A reference to the changed tNN object with the data added.
Note: tNN objects store all variable data in an environment which
enables us to update partial data without copying the whole object. Assignment
will not create a copy! Use the provided method \code{copy()}.
}
\description{
Cluster new data into an existing tNN object.
}
\details{
\code{cluster()} implements tNN clustering The dissimilarity between
	the new observation and the centers of the clusters is calculated. The
	new observation is assigned to the closest cluster if the dissimilarity
	value is smaller than the threshold (for the state).  If no such state
	exists, a new state is created for the observation. This simple
	clustering algorithm is called nearest neighbor threshold nearest
	neighbor (threshold NN).

	\code{NA}s are handled in the data by using only the other
	dimensions if the data for dissimilarity computation
	(see package~\pkg{proxy}).

	The clusters which the data points in the last \code{cluster()}
	operation where assigned to can be retrieved using the method
	\code{last_clustering()}.
}
\examples{
## load EMMTraffic data
data(EMMTraffic)

## create empty clustering
tnn <- tNN(threshold = 0.2, measure = "eJaccard")
tnn

## cluster some data
cluster(tnn, EMMTraffic)
tnn

## what clusters were the data points assigned to?
last_clustering(tnn)

## plot the clustering as a scatterplot matrix of the cluster centers
plot(tnn)
}
\seealso{
Class \code{\linkS4class{tNN}},
	\code{\link{fade}} and \code{\link[proxy]{dist}}
	in \pkg{proxy}.
}
\keyword{cluster}
\keyword{models}
