% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/EMM.R
\docType{class}
\name{EMM-class}
\alias{EMM-class}
\alias{show,EMM-method}
\alias{size}
\alias{size,EMM-method}
\alias{copy}
\alias{copy,EMM-method}
\alias{EMM}
\alias{object.size,EMM-method}
\title{Class "EMM"}
\usage{
EMM(
  threshold = 0.2,
  measure = "euclidean",
  distFun = NULL,
  centroids = identical(tolower(measure), "euclidean"),
  lambda = 0,
  data = NULL
)
}
\arguments{
\item{threshold}{Object of class \code{"numeric"} with the dissimilarity threshold used by the
clustering algorithm for assigning a new observation to existing clusters.}

\item{measure}{Object of class \code{"character"} containing the name of the dissimilarity
measure used (see \code{dist} in \pkg{proxy} for available measures).}

\item{distFun}{Specify a function passed on as method to \code{dist} in \pkg{proxy}
(see \code{dist} in \pkg{proxy}).
The character string passed on as \code{measure} will be used as the
measure's name.}

\item{centroids}{Object of class \code{"logical"} indicating if centroids are used for clusters.
If \code{FALSE}, pseudo medians (first observation of a cluster) are used to
represent a cluster.}

\item{lambda}{Object of class \code{"numeric"} specifying the
rate for fading.}

\item{data}{Initial data to build the EMM.
This just calls \code{build} on the new EMM.}
}
\value{
An object of class \code{"EMM"}.
}
\description{
This class represents the extensible Markov Model.	It consists
of a simple data stream clustering algorithm (class \code{"tNN"}) and
a temporal layer (class \code{"TRACDS"}).

Create a new object of class \code{"EMM"}.
}
\section{Objects from the Class}{

Objects can be created using the creator function \code{EMM} or by
directly calling \code{new("EMM", ...)}. Most slots for the extended
classes can be used as parameters for \code{EMM}.
}

\section{Slots}{

The slots are described in corresponding the extended classes
	(see section Extends).
}

\section{Extends}{

Class \code{"\linkS4class{tNN}"}, directly.
Class \code{"\linkS4class{TRACDS}"}, directly.
}

\section{Methods}{

\describe{
    \item{copy}{\code{signature(x = "EMM")}: Make a copy of the EMM object.
    Making explicit copies is necessary since the subclasses store
    information in environments which are not copied for regular
    assignements.}
    \item{size}{\code{signature(x = "EMM")}: Returns the size of
    the EMM (number of clusters/states).}


















	 }
}

\examples{
## load EMMTraffic data
data(EMMTraffic)

## create empty EMM
emm <- EMM(threshold=0.2, measure="eJaccard", lambda=.1)
emm

## cluster some data
build(emm, EMMTraffic)
emm

## what clusters were the data points assigned to?
last_clustering(emm)

## plot the clustering as a graph
plot(emm)
}
\references{
M.H. Dunham, Y. Meng, J. Huang (2004): Extensible Markov Model, In:
\emph{ICDM '04: Proceedings of the Fourth IEEE International
Conference on Data Mining,} pp. 371--374.
}
\seealso{
\code{\link{build}},
	\code{\link{fade}},
	\code{\link{merge_clusters}},
	\code{\link{plot}},
	\code{\link{prune}},
	\code{\link{rare_clusters}},
	\code{\link{rare_transitions}},
	\code{\link{remove_clusters}},
	\code{\link{remove_transitions}},
	\code{\link{remove_selftransitions}},
	\code{\link{recluster}}, and
	\code{\link{score}}.

\code{\link{EMM-class}}
}
\keyword{classes}
\keyword{models}
