% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EMMTraffic}
\alias{EMMTraffic}
\title{Hypothetical Traffic Data Set for EMM}
\format{
A matrix with 12 observations (rows).
}
\usage{
data(EMMTraffic)
}
\description{
Each observation in this hypothetical data set is a vector of
seven values obtained from sensors located at specific
points on roads. Each sensor collects a count of the
number of vehicles which have crossed this sensor in
the preceding time interval.
}
\examples{
data(EMMTraffic)
EMMTraffic
}
\references{
M.H. Dunham, Y. Meng, J. Huang (2004): Extensible Markov Model, In:
\emph{ICDM '04: Proceedings of the Fourth IEEE International
Conference on Data Mining,} pp. 371--374.
}
\keyword{datasets}
