% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/prune.R
\name{prune}
\alias{prune}
\alias{prune,EMM-method}
\alias{rare_clusters}
\alias{rare_clusters,tNN-method}
\alias{rare_transitions}
\alias{rare_transitions,TRACDS-method}
\title{Prune States and/or Transitions}
\usage{
rare_transitions(x, count_threshold, ...)

rare_clusters(x, count_threshold, ...)

\S4method{prune}{EMM}(
  x,
  count_threshold,
  clusters = TRUE,
  transitions = FALSE,
  copy = TRUE,
  compact = TRUE
)
}
\arguments{
\item{x}{an object of class \code{"EMM"}}

\item{count_threshold}{all states/edges with a count of less or equal to the
threshold are removed from the model.}

\item{...}{further arguments (currently not used).}

\item{clusters}{logical; prune clusters?}

\item{transitions}{logical; prune transitions?}

\item{copy}{logical; make a copy of x before reclustering? Otherwise the function will change \code{x}!}

\item{compact}{logical; tries make the data structure used for the temporal model more compact after pruning.}
}
\value{
\code{prune} returns invisibly an object of class \code{EMM}.
    If \code{copy=FALSE} then it returns a reference to the changes
    object passed as \code{x}.

    \code{rare_clusters} returns a vector of names of rare clusters.

    \code{rare_transitions} returns a data.frame of rare transitions.
}
\description{
Simplifies an EMM and/or the clustering by removing all
clusters/states and/or transitions
which have a count of equal or smaller than a given threshold.
}
\examples{
data("EMMTraffic")

## For the example we use a very high learning rate
emm_l <- EMM(threshold=0.2, measure="eJaccard", lambda = 1)
build(emm_l, EMMTraffic)

## show state counts and transition counts
cluster_counts(emm_l)
transition_matrix(emm_l, type="counts")

## rare state/transitions
rare_clusters(emm_l, count_threshold=0.1)
rare_transitions(emm_l, count_threshold=0.1)

## remove all states with a threshold of 0.1
emm_lr <- prune(emm_l, count_threshold=0.1)

## compare graphs
op <- par(mfrow = c(1, 2), pty = "m")
plot(emm_l, main = "EMM with high learning rate")
plot(emm_lr, main = "Simplified EMM")
par(op)
}
\seealso{
\code{\link{remove_transitions}},
\code{\link{remove_clusters}},
\code{\link{compact}}
}
\keyword{manip}
