% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove}
\alias{remove}
\alias{remove_clusters}
\alias{remove_clusters,EMM,character-method}
\alias{remove_transitions}
\alias{remove_transitions,EMM,character,character-method}
\alias{remove_transitions,EMM,matrix,missing-method}
\alias{remove_selftransitions}
\alias{remove_selftransitions,EMM-method}
\title{Remove States/Clusters or Transitions from an EMM}
\usage{
remove_clusters(x, to_remove, copy = TRUE)
remove_transitions(x, from, to,copy = TRUE)
remove_selftransitions(x, copy = TRUE)
}
\arguments{
\item{x}{an \code{EMM} object.}

\item{to_remove}{Names of states/clusters to remove.}

\item{copy}{logical; make a copy of x before reclustering? Otherwise the function will change \code{x}!}

\item{from, to}{Names of states for removing transitions. If \code{to} is
missing \code{from} has to contain a matrix with two columns
(from and to state names).}
}
\value{
Returns a EMM with removed states/transitions.
If \code{copy=FALSE} a reference to the object \code{x}
with the states/transistions removed is returned.
}
\description{
Remove states/clusters or transitions from an EMM.
}
\details{
\code{remove_selftransitions} removes the transitions from each state to itself.
}
\examples{
data("EMMTraffic")
emm <- EMM(measure="eJaccard", threshold=0.2)
emm <- build(emm, EMMTraffic)

## remove state 3
emm_rs3 <- remove_clusters(emm, "3")

## remove transition 5->2
emm_rt52 <- remove_transitions(emm, "5", "2")

## compare EMMs
op <- par(mfrow = c(2, 2), pty = "m")
plot(emm, method = "igraph", main = "original EMM")
plot(emm_rs3, method = "igraph", main = "state 3 removed")
plot(emm_rt52, method = "igraph", main = "transition 5->2 removed")
par(op)
}
\keyword{manip}
