% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact.R, R/reset.R, R/update.R
\name{update}
\alias{update}
\alias{update,TRACDS-method}
\alias{reset}
\alias{reset,TRACDS-method}
\alias{compact}
\alias{compact,TRACDS-method}
\title{Update a TRACDS temporal structure with new state assignements}
\usage{
\S4method{compact}{TRACDS}(x)

\S4method{reset}{TRACDS}(x)

\S4method{update}{TRACDS}(object, newdata, verbose = FALSE, ...)
}
\arguments{
\item{x, object}{a \code{TRACDS} object. Note that this function changes the original object!}

\item{newdata}{a vector with a state assignemnt sequence (typically
produced by clustering).}

\item{verbose}{logical; verbose output?}

\item{...}{further arguments.}
}
\value{
A reference to the changed TRACDS object with the data added.
Note: EMM objects store all variable data in an environment which
enables us to update partial data without copying the whole object. Assignment
will not create a copy! Use the provided method \code{copy()}.
}
\description{
Add a sequence of new state assignments to a TRACDS object.
}
\details{
\code{update()}
    adds a new state assignemnt sequenc to the TRACDS object by increasing the
    transition counts and, if needed, creating new states.

    \code{reset()} resets the current state to \code{NA} for reading in a
    new sequence. An \code{NA} in \code{newdata}
    also resets the current state.

    \code{compact()} reduces the size (memory) used to store the temporal
    transition matrix.
}
\examples{
## create an empty TRACDS object
tracds <- TRACDS()
tracds

## update with an cluster assignment sequence
update(tracds, c(1,2,5,5,2))
tracds

plot(tracds)
}
\seealso{
Class \code{\linkS4class{TRACDS}},
	\code{\link{fade}}.
}
\keyword{models}
